+++ src/controllers/TunnelController.js
@@ -1,7 +1,9 @@
+const Host = require('../models/Host');
 const Tunnel = require('../models/Tunnel');
 const TunnelController = {};
 
-TunnelController.createTunnel = async function (name, description, hostID, tenantID, hostConnectPort, wgListeningPort, type) {
+TunnelController.createTunnel = async function (name, description, hostID, tenantID, hostConnectPort, type) {
+    const wgListeningPort = await Host.registerWGPort(hostID);
     const tunnel = await Tunnel.create({
         name, description, hostID, tenantID, hostConnectPort, wgListeningPort, type
     });
@@ -22,9 +24,18 @@ TunnelController.getTunnel = async function(id) {
 }
 
 TunnelController.deleteTunnel = async function(id) {
-    await Tunnel.deleteOne({
+    const tunnel = await Tunnel.findOne({
         _id: id
-    })
+    }).select('hostID', 'wgListeningPort');
+
+    if(tunnel){
+        await Host.unregisterWGPort(tunnel["hostID"], tunnel["wgListeningPort"]);
+
+        await Tunnel.deleteOne({
+            _id: id
+        })
+    }
+
 }
 
 TunnelController.getTunnelsOfTenant = function(tenantID) {
